+++ src/components/BarChart/charts.stories.mdx
@@ -1,18 +1,18 @@
 import { Meta, Story } from "@storybook/addon-docs";
-import { BarChart, LineChart, PieChart } from "../../components";
+import { BarChart } from "./";
 import { BrowserRouter } from "react-router-dom";
 import { Box } from "@mui/material";
 
 <Meta
-  title="Components/Charts"
+  title="Components/BarChart"
   parameters={{
     viewMode: "docs",
   }}
 />
 
-# Charts
+# BarChart
 
-<Story name="Charts">
+<Story name="BarChart">
   <BrowserRouter>
     <Box sx={{ height: "400px", width: "600px" }}>
       <BarChart
