+++ src/components/Header/header.stories.mdx
@@ -11,7 +11,7 @@ import {
   Stack,
   Box,
 } from "@mui/material";
-import { Header } from "../../components";
+import { Header } from "./";
 import { BrowserRouter } from "react-router-dom";
 
 <Meta
