+++ src/components/Spinner/spinner.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story } from "@storybook/addon-docs";
-import { Spinner } from "../../components";
+import { Spinner } from "./";
 import { BrowserRouter } from "react-router-dom";
 
 <Meta
