+++ src/components/TextField/input.stories.mdx
@@ -11,7 +11,7 @@ import {
   Stack,
   Box,
 } from "@mui/material";
-import { TextField } from "../../components";
+import { TextField } from "./";
 import { BrowserRouter } from "react-router-dom";
 
 <Meta
