+++ src/components/Theme/Colors.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from "@storybook/addon-docs";
 import { Box, Grid, Typography } from "@mui/material";
-import ThemeProvider from "../../components/Theme/ThemeProvider";
+import ThemeProvider from "./ThemeProvider";
 import { ScopedCssBaseline } from "@mui/material";
 
 <style>{`
